move status task to its own class
